@@ -3,7 +3,7 @@ package simapp
 import (
 	"encoding/json"
 	"io"
-	"io/ioutil"
+	"os"
 
 	"github.com/Pylons-tech/pylons/app"
 	tmjson "github.com/tendermint/tendermint/libs/json"
@@ -18,7 +18,7 @@ import (
 // AppStateFromGenesisFileFn util function to generate the genesis AppState
 // from a genesis.json file.
 func AppStateFromGenesisFileFn(r io.Reader, cdc codec.JSONCodec, genesisFile string) (tmtypes.GenesisDoc, []simtypes.Account) {
-	bytes, err := ioutil.ReadFile(genesisFile)
+	bytes, err := os.ReadFile(genesisFile)
 	if err != nil {
 		panic(err)
 	}
