@@ -17,7 +17,7 @@ import (
 
 var Out io.Writer = os.Stdout // modified during testing
 
-// group 1: (whole raw string tokens encapsulated '''like this'')
+// group 1: (whole raw string tokens encapsulated like this with three ```)
 // group 2: (tokens not containing whitespace, separated by whitespace)
 // group 3: (whatever is in front of the first whitespace)
 var gadgetParamParseRegex = regexp.MustCompile(`(\s'''.*''')|(\s.\S*)|(\S*)`)
